Fix for PowerSender - Remove all previous log files when execution starts

+++ ansible/provisioning/scripts/PowerSender/src/utils/MyUtils.py
@@ -9,6 +9,9 @@ class MyUtils:
             os.makedirs(path)
 
     @staticmethod
-    def clean_log_file(log_file):
-        if os.path.isfile(log_file):
-            os.remove(log_file)
+    def clean_log_files(log_dir):
+        for filename in os.listdir(log_dir):
+            if ".log" in filename:
+                log_file = os.path.join(log_dir, filename)
+                if os.path.isfile(log_file):
+                    os.remove(log_file)
